Created the first ideas for a broker interface and made some improvements for the algorithm related stuff
One of the biggest changes is the new brokers folder with broker related stuff. Even if this project currently is called `algorithm-utils`, I think it's best to combine it with the broker interface, since a broker interface also needs to know about derivatives, orders, ...
I will rename the project in the next few commits.
The broker interface is far from finished. Currently I just wrote down some incomplete ideas. I will start just writing a broker interface for the comdirect  since I have a deposit there and don't know any other apis yet. The wide range of other functionality will be implemented when needed.

Also I plan to rethink the way of error handling. I already started coding some aspects of it but it's not finished yet. My decision is to have one central Error struct with a lot of different ErrorKind Traits. By this programmers don't have to check hundreds of error kinds when getting back an error.

When it comes to algorithms I made some functions const (what hopefully brings a performance improvement). I also moved some information about the algorithm out of the AlgorithmInterface into the export_algorithm! macro. The last important change was to move the load function out of the Algorithms struct into the Algorithm struct. This is just for more expressive code since the load function was basically a constructor. Now it's not just basically a constructor any more.

The next steps will be to improve some ideas related to broker and bank related stuff.