Added Deposit and Position functionality
I just added the Deposit (deposit.rs) functionality. The Deposit struct is meant to be an abstraction of a Deposit that allows to represent important information like orders, positions or the currently available cash. It will be a struct used by the BrokerInterfaces to be able to store this important data properly. It's job is not to hold the complete Deposit state and information.

Also I added some functionality to Position (position.rs). It now holds more and more important information.